huggingface_space/requirements.txt
gradio>=6.9.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
huggingface_hub>=0.20.0
